Updated the plate well sampling to use the dp solution
